feat: use secrets overview as app home
